@@ -1,29 +1,28 @@
-module github.com/gsakun/sonarqube-client-go
+module github.com/boxboxjason/sonarqube-client-go
 
-go 1.21.13
+go 1.25.5
 
 require (
-	github.com/dave/jennifer v1.2.0
+	github.com/dave/jennifer v1.7.1
 	github.com/davecgh/go-spew v1.1.1
-	github.com/fatih/color v1.7.0
-	github.com/google/go-querystring v1.0.0
-	github.com/json-iterator/go v1.1.5
+	github.com/fatih/color v1.18.0
+	github.com/google/go-querystring v1.1.0
+	github.com/json-iterator/go v1.1.12
 	github.com/magicsong/color-glog v0.0.1
-	github.com/magicsong/generate-go-for-sonarqube v0.0.0-20181119114409-6374ba20fed2
 	github.com/onsi/ginkgo v1.6.0
 	github.com/onsi/gomega v1.4.2
-	gopkg.in/yaml.v2 v2.2.1
+	gopkg.in/yaml.v2 v2.4.0
 )
 
 require (
 	github.com/hpcloud/tail v1.0.0 // indirect
-	github.com/mattn/go-colorable v0.0.9 // indirect
-	github.com/mattn/go-isatty v0.0.4 // indirect
+	github.com/mattn/go-colorable v0.1.14 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
-	github.com/modern-go/reflect2 v0.0.0-20180701023420-4b7aa43c6742 // indirect
+	github.com/modern-go/reflect2 v1.0.2 // indirect
 	github.com/stretchr/testify v1.10.0 // indirect
 	golang.org/x/net v0.0.0-20181108082009-03003ca0c849 // indirect
-	golang.org/x/sys v0.0.0-20181107165924-66b7b1311ac8 // indirect
+	golang.org/x/sys v0.39.0 // indirect
 	golang.org/x/text v0.3.0 // indirect
 	gopkg.in/fsnotify.v1 v1.4.7 // indirect
 	gopkg.in/tomb.v1 v1.0.0-20141024135613-dd632973f1e7 // indirect
