@@ -8,11 +8,11 @@ import (
 	"strings"
 
 	. "github.com/dave/jennifer/jen"
-	"github.com/magicsong/color-glog"
+	glog "github.com/magicsong/color-glog"
 
-	"github.com/magicsong/generate-go-for-sonarqube/pkg/api"
-	"github.com/magicsong/generate-go-for-sonarqube/pkg/response"
-	"github.com/magicsong/generate-go-for-sonarqube/pkg/util/strcase"
+	"github.com/boxboxjason/sonarqube-client-go/pkg/api"
+	"github.com/boxboxjason/sonarqube-client-go/pkg/response"
+	"github.com/boxboxjason/sonarqube-client-go/pkg/util/strcase"
 )
 
 var (
@@ -67,23 +67,23 @@ func prepare(pkgName, workingDir, endpoint, username, password string) error {
 	Endpoint = endpoint
 	Username = username
 	Password = password
-	_ = os.Mkdir(workingDir, os.ModeDir)
-	_ = os.Mkdir("integration_testing", os.ModeDir)
-	_ = os.Mkdir("integration_testing/run", os.ModeDir)
-	wd, err := os.Getwd()
+	_ = os.Mkdir(workingDir, 0755)
+	_ = os.Mkdir("integration_testing", 0755)
+	_ = os.Mkdir("integration_testing/run", 0755)
+
+	CurrentRepo, err := GetCurrentModulePath()
 	if err != nil {
-		glog.Errorln("FAILED to get current working directory")
+		glog.Errorln("Failed to get the module path from go.mod")
 		return err
 	}
-	CurrentRepo, err = GetCurrentRepo(workingDir, wd)
-	if err != nil {
-		glog.Errorln("Failed to get the full package name")
-		return err
+	// Append working directory to module path if it's not the root
+	if workingDir != "." && workingDir != "" {
+		CurrentRepo = CurrentRepo + "/" + workingDir
 	}
 	//setup validation file
 	validation = NewFile(pkgName)
-	validation.HeaderComment("[TODO] Code generated by \"github.com/magicsong/generate-go-for-sonarqube\", you should edit some validation logic here")
-	validation.ImportName("github.com/magicsong/generate-go-for-sonarqube/pkg/validation", "validation")
+	validation.HeaderComment("[TODO] Code generated by \"github.com/boxboxjason/sonarqube-client-go\", you should edit some validation logic here")
+	validation.ImportName("github.com/boxboxjason/sonarqube-client-go/pkg/validation", "validation")
 	//set up can-be-run test file
 	fTestRun = NewFile("main")
 	fTestRun.ImportAlias(CurrentRepo, ".")
@@ -119,6 +119,11 @@ func Build(pkgName, workingDir, endpoint, username, password string, apidoc *api
 	}
 	for _, item := range apidoc.WebServices {
 		name := item.Path[4:]
+		// Skip empty service names
+		if name == "" {
+			glog.V(1).Infof("Skipping service with empty name: %s\n", item.Path)
+			continue
+		}
 		contin := false
 		for _, word := range DeprecatedWords {
 			if strings.Contains(strings.ToLower(item.Description), word) {
@@ -316,10 +321,43 @@ func GenerateGoContent(packageName string, service *api.WebService) (f *File, er
 	fetcher := response.NewExampleFetcher(Endpoint, Username, Password)
 	examples, err := fetcher.GetResponseExample(service)
 	if err != nil {
-		glog.Errorf("cannot fetch examples of <%s>", service.Path)
-		return nil, err
+		glog.Warningf("cannot fetch examples of <%s>: %v (continuing without examples)", service.Path, err)
+		examples = make([]*response.WebservicesResponseExampleResp, 0)
+		err = nil
+		// When examples can't be fetched, we need to set default response types
+		// for all actions so the code generator knows how to handle them
+		for i := range service.Actions {
+			// If ResponseType is not already set, we need to determine a default
+			if service.Actions[i].ResponseType == "" {
+				if service.Actions[i].Post {
+					// POST requests without explicit response type typically return no body
+					service.Actions[i].ResponseType = "no-content"
+				} else {
+					// GET requests default to JSON
+					service.Actions[i].ResponseType = "json"
+					// Only generate the struct if it's supposed to have a response example
+					if service.Actions[i].HasResponseExample {
+						respName := strcase.ToCamel(name + "_" + service.Actions[i].Key + "Object")
+						f.Commentf("[TODO] cannot fetch response example for <%s>, struct needs to be filled manually", service.Actions[i].Key)
+						f.Type().Id(respName).Struct().Line()
+					}
+				}
+			}
+		}
 	}
 	for _, exam := range examples {
+		if exam.Format == "proto" {
+			glog.V(1).Infof("The response of action <%s> for api <%s> is proto, using []byte", exam.Name, name)
+			// Proto responses should be treated as binary data
+			// Find the corresponding action and set its response type
+			for i := range service.Actions {
+				if service.Actions[i].Key == exam.Name {
+					service.Actions[i].ResponseType = "proto"
+					break
+				}
+			}
+			continue
+		}
 		if exam.Format != "json" {
 			glog.V(1).Infof("The response of action <%s> for api <%s> is %s, not json", exam.Name, name, exam.Format)
 			continue
@@ -353,7 +391,7 @@ func detectDeprecatedField(field *api.Param) bool {
 	return false
 }
 
-//GenerateServiceActionContent generate code of each service,include api method and related structs
+// GenerateServiceActionContent generate code of each service,include api method and related structs
 func GenerateServiceActionContent(serviceName string, action *api.Action) *Statement {
 	c := Line()
 	hasOption := true
@@ -398,6 +436,13 @@ func GenerateServiceActionContent(serviceName string, action *api.Action) *State
 		case "txt", "log", "svg", "xml":
 			g.Id("v").Op("*").String()
 			respName = "string"
+		case "proto":
+			// Protocol Buffer binary data
+			g.Id("v").Op("[]").Byte()
+			respName = "[]byte"
+		case "no-content", "":
+			// No response body expected
+			noResp = true
 		default:
 			if method == "GET" {
 				g.Id("v").Op("*").Id(respName)
@@ -425,6 +470,14 @@ func GenerateServiceActionContent(serviceName string, action *api.Action) *State
 		if noResp {
 			g.List(Id("resp"), Err()).Op("=").Id("s").Dot("client").Dot("Do").Call(Id("req"), Nil())
 			ErrorHandlerHelper(g)
+		} else if action.ResponseType == "proto" {
+			// For proto responses, don't use New() since v is []byte not a pointer
+			g.List(Id("resp"), Err()).Op("=").Id("s").Dot("client").Dot("Do").Call(Id("req"), Op("&").Id("v"))
+			g.If(
+				Err().Op("!=").Nil(),
+			).Block(
+				Return().List(Nil(), Id("resp"), Err()),
+			)
 		} else {
 			g.Id("v").Op("=").New(Id(respName))
 			g.List(Id("resp"), Err()).Op("=").Id("s").Dot("client").Dot("Do").Call(Id("req"), Id("v"))
@@ -447,17 +500,27 @@ func ErrorHandlerHelper(g *Group) {
 	)
 }
 
-func GetCurrentRepo(workingDir, currentDir string) (string, error) {
-	gopath := os.Getenv("GOPATH")
-	if gopath == "" {
-		return "", errors.New("$GOPATH is not set correctly")
+// GetCurrentModulePath reads the module path from go.mod file.
+// This replaces the legacy GOPATH-based approach.
+func GetCurrentModulePath() (string, error) {
+	// Read go.mod file
+	data, err := ioutil.ReadFile("go.mod")
+	if err != nil {
+		return "", fmt.Errorf("failed to read go.mod: %w (make sure you're in a Go module directory)", err)
 	}
-	gosrc := gopath + "/src"
-	if currentDir[:len(gosrc)] != gosrc {
-		return "", errors.New("current location is not in $GOPATH/src")
+
+	// Parse module path from go.mod
+	lines := strings.Split(string(data), "\n")
+	for _, line := range lines {
+		line = strings.TrimSpace(line)
+		if strings.HasPrefix(line, "module ") {
+			modulePath := strings.TrimSpace(strings.TrimPrefix(line, "module"))
+			if modulePath == "" {
+				return "", errors.New("module path is empty in go.mod")
+			}
+			return modulePath, nil
+		}
 	}
-	if len(currentDir) == len(gosrc) {
-		return "", errors.New("current location should be in subdir of $GOPATH/src")
-	}
-	return currentDir[len(gosrc)+1:] + "/" + workingDir, nil
+
+	return "", errors.New("module directive not found in go.mod")
 }
