@@ -1,5 +1,5 @@
 // Get information about a component (file, directory, project, ...) and its ancestors or descendants. Update a project or module key.
-package sonar
+package sonargo
 
 import "net/http"
 
@@ -54,19 +54,22 @@ type ComponentsSearchProjectsObject struct {
 	Paging     ComponentsSearchProjectsObject_sub4   `json:"paging,omitempty"`
 }
 
+type ComponentsSearchProjectsObject_sub1 struct {
+	AiCodeAssurance    string   `json:"aiCodeAssurance,omitempty"`
+	ContainsAiCode     bool     `json:"containsAiCode,omitempty"`
+	IsAiCodeFixEnabled bool     `json:"isAiCodeFixEnabled,omitempty"`
+	IsFavorite         bool     `json:"isFavorite,omitempty"`
+	Key                string   `json:"key,omitempty"`
+	Name               string   `json:"name,omitempty"`
+	Qualifier          string   `json:"qualifier,omitempty"`
+	Tags               []string `json:"tags,omitempty"`
+	UUID               string   `json:"uuid,omitempty"`
+	Visibility         string   `json:"visibility,omitempty"`
+}
+
 type ComponentsSearchProjectsObject_sub2 struct {
 	Count int64  `json:"count,omitempty"`
 	Val   string `json:"val,omitempty"`
-}
-
-type ComponentsSearchProjectsObject_sub1 struct {
-	IsFavorite    bool     `json:"isFavorite,omitempty"`
-	Key           string   `json:"key,omitempty"`
-	Name          string   `json:"name,omitempty"`
-	NeedIssueSync bool     `json:"needIssueSync,omitempty"`
-	Qualifier     string   `json:"qualifier,omitempty"`
-	Tags          []string `json:"tags,omitempty"`
-	Visibility    string   `json:"visibility,omitempty"`
 }
 
 type ComponentsSearchProjectsObject_sub4 struct {
@@ -183,7 +186,7 @@ func (s *ComponentsService) App(opt *ComponentsAppOption) (v *ComponentsAppObjec
 type ComponentsSearchOption struct {
 	P          string `url:"p,omitempty"`          // Description:"1-based page number",ExampleValue:"42"
 	Ps         string `url:"ps,omitempty"`         // Description:"Page size. Must be greater than 0 and less or equal than 500",ExampleValue:"20"
-	Q          string `url:"q,omitempty"`          // Description:"Limit search to: <ul><li>component names that contain the supplied string</li><li>component keys that are exactly the same as the supplied string</li></ul>",ExampleValue:"sonar"
+	Q          string `url:"q,omitempty"`          // Description:"Limit search to: <ul><li>component names that contain the supplied string</li><li>component keys that are exactly the same as the supplied string</li></ul><br>The value length of the param must be between 2 and 15 (inclusive) characters. In case longer value is provided it will be truncated.",ExampleValue:"sonar"
 	Qualifiers string `url:"qualifiers,omitempty"` // Description:"Comma-separated list of component qualifiers. Filter the results with the specified qualifiers. Possible values are:<ul><li>TRK - Projects</li></ul>",ExampleValue:""
 }
 
@@ -209,10 +212,10 @@ type ComponentsSearchProjectsOption struct {
 	Asc    string `url:"asc,omitempty"`    // Description:"Ascending sort",ExampleValue:""
 	F      string `url:"f,omitempty"`      // Description:"Comma-separated list of the fields to be returned in response",ExampleValue:""
 	Facets string `url:"facets,omitempty"` // Description:"Comma-separated list of the facets to be computed. No facet is computed by default.",ExampleValue:""
-	Filter string `url:"filter,omitempty"` // Description:"Filter of projects on name, key, measure value, quality gate, language, tag or whether a project is a favorite or not.<br>The filter must be encoded to form a valid URL (for example '=' must be replaced by '%3D').<br>Examples of use:<ul> <li>to filter my favorite projects with a failed quality gate and a coverage greater than or equals to 60% and a coverage strictly lower than 80%:<br>   <code>filter="alert_status = ERROR and isFavorite and coverage >= 60 and coverage < 80"</code></li> <li>to filter projects with a reliability, security and maintainability rating equals or worse than B:<br>   <code>filter="reliability_rating>=2 and security_rating>=2 and sqale_rating>=2"</code></li> <li>to filter projects without duplication data:<br>   <code>filter="duplicated_lines_density = NO_DATA"</code></li></ul>To filter on project name or key, use the 'query' keyword, for instance : <code>filter='query = "Sonar"'</code>.<br><br>To filter on a numeric metric, provide the metric key.<br>These are the supported metric keys:<br><ul><li>alert_status</li><li>coverage</li><li>duplicated_lines_density</li><li>lines</li><li>ncloc</li><li>ncloc_language_distribution</li><li>new_coverage</li><li>new_duplicated_lines_density</li><li>new_lines</li><li>new_maintainability_rating</li><li>new_reliability_rating</li><li>new_security_hotspots_reviewed</li><li>new_security_rating</li><li>new_security_review_rating</li><li>reliability_rating</li><li>security_hotspots_reviewed</li><li>security_rating</li><li>security_review_rating</li><li>sqale_rating</li></ul><br>To filter on a rating, provide the corresponding metric key (ex: reliability_rating for reliability rating).<br>The possible values are:<ul> <li>'1' for rating A</li> <li>'2' for rating B</li> <li>'3' for rating C</li> <li>'4' for rating D</li> <li>'5' for rating E</li></ul>To filter on a Quality Gate status use the metric key 'alert_status'. Only the '=' operator can be used.<br>The possible values are:<ul> <li>'OK' for Passed</li> <li>'WARN' for Warning</li> <li>'ERROR' for Failed</li></ul>To filter on language keys use the language key: <ul> <li>to filter on a single language you can use 'language = java'</li> <li>to filter on several languages you must use 'language IN (java, js)'</li></ul>Use the WS api/languages/list to find the key of a language.<br> To filter on tags use the 'tags' keyword:<ul> <li>to filter on one tag you can use <code>tags = finance</code></li> <li>to filter on several tags you must use <code>tags in (offshore, java)</code></li></ul>To filter on a qualifier use key 'qualifier'. Only the '=' operator can be used.<br>The possible values are:<ul> <li>TRK - for projects</li> <li>APP - for applications</li></ul>",ExampleValue:""
+	Filter string `url:"filter,omitempty"` // Description:"Filter of projects on name, key, measure value, quality gate, language, tag or whether a project is a favorite or not.<br>The filter must be encoded to form a valid URL (for example '=' must be replaced by '%3D').<br>Examples of use:<ul> <li>to filter my favorite projects with a failed quality gate and a coverage greater than or equals to 60% and a coverage strictly lower than 80%:<br>   <code>filter="alert_status = ERROR and isFavorite and coverage >= 60 and coverage < 80"</code></li> <li>to filter projects with a reliability, security and maintainability rating equals or worse than B:<br>   <code>filter="reliability_rating>=2 and security_rating>=2 and sqale_rating>=2"</code></li> <li>to filter projects without duplication data:<br>   <code>filter="duplicated_lines_density = NO_DATA"</code></li></ul>To filter on project name or key, use the 'query' keyword, for instance : <code>filter='query = "Sonar"'</code>.<br><br>To filter on a numeric metric, provide the metric key.<br>These are the supported metric keys:<br><ul><li>alert_status</li><li>coverage</li><li>duplicated_lines_density</li><li>lines</li><li>ncloc</li><li>ncloc_language_distribution</li><li>new_coverage</li><li>new_duplicated_lines_density</li><li>new_lines</li><li>new_maintainability_rating</li><li>new_reliability_rating</li><li>new_security_hotspots_reviewed</li><li>new_security_rating</li><li>new_security_review_rating</li><li>new_software_quality_maintainability_rating</li><li>new_software_quality_reliability_rating</li><li>new_software_quality_security_rating</li><li>reliability_rating</li><li>security_hotspots_reviewed</li><li>security_rating</li><li>security_review_rating</li><li>software_quality_maintainability_rating</li><li>software_quality_reliability_rating</li><li>software_quality_security_rating</li><li>sqale_rating</li></ul><br>To filter on a rating, provide the corresponding metric key (ex: reliability_rating for reliability rating).<br>The possible values are:<ul> <li>'1' for rating A</li> <li>'2' for rating B</li> <li>'3' for rating C</li> <li>'4' for rating D</li> <li>'5' for rating E</li></ul>To filter on a Quality Gate status use the metric key 'alert_status'. Only the '=' operator can be used.<br>The possible values are:<ul> <li>'OK' for Passed</li> <li>'WARN' for Warning</li> <li>'ERROR' for Failed</li></ul>To filter on languages use the 'languages' keyword: <ul> <li>to filter on a single language you can use 'languages = java'</li> <li>to filter on several languages you must use 'languages IN (java, js)'</li></ul>Use the WS api/languages/list to find the key of a language.<br> To filter on tags use the 'tags' keyword:<ul> <li>to filter on one tag you can use <code>tags = finance</code></li> <li>to filter on several tags you must use <code>tags in (offshore, java)</code></li></ul>To filter on a qualifier use key 'qualifier'. Only the '=' operator can be used.<br>The possible values are:<ul> <li>TRK - for projects</li> <li>APP - for applications</li></ul>",ExampleValue:""
 	P      string `url:"p,omitempty"`      // Description:"1-based page number",ExampleValue:"42"
 	Ps     string `url:"ps,omitempty"`     // Description:"Page size. Must be greater than 0 and less or equal than 500",ExampleValue:"20"
-	S      string `url:"s,omitempty"`      // Description:"Sort projects by numeric metric key, quality gate status (using 'alert_status'), last analysis date (using 'analysisDate'), or by project name.",ExampleValue:""
+	S      string `url:"s,omitempty"`      // Description:"Sort projects by numeric metric key, quality gate status (using 'alert_status'), last analysis date (using 'analysisDate'), project name or creationDate (using 'filter').",ExampleValue:""
 }
 
 // SearchProjects Search for projects
@@ -289,7 +292,7 @@ type ComponentsTreeOption struct {
 	Ps          string `url:"ps,omitempty"`          // Description:"Page size. Must be greater than 0 and less or equal than 500",ExampleValue:"20"
 	PullRequest string `url:"pullRequest,omitempty"` // Description:"Pull request id. Not available in the community edition.",ExampleValue:"5461"
 	Q           string `url:"q,omitempty"`           // Description:"Limit search to: <ul><li>component names that contain the supplied string</li><li>component keys that are exactly the same as the supplied string</li></ul>",ExampleValue:"FILE_NAM"
-	Qualifiers  string `url:"qualifiers,omitempty"`  // Description:"Comma-separated list of component qualifiers. Filter the results with the specified qualifiers. Possible values are:<ul><li>BRC - no description available</li><li>UTS - Test Files</li><li>FIL - Files</li><li>DIR - Directories</li><li>TRK - Projects</li></ul>",ExampleValue:""
+	Qualifiers  string `url:"qualifiers,omitempty"`  // Description:"Comma-separated list of component qualifiers. Filter the results with the specified qualifiers. Possible values are:<ul><li>UTS - Test Files</li><li>FIL - Files</li><li>DIR - Directories</li><li>TRK - Projects</li></ul>",ExampleValue:""
 	S           string `url:"s,omitempty"`           // Description:"Comma-separated list of sort fields",ExampleValue:"name, path"
 	Strategy    string `url:"strategy,omitempty"`    // Description:"Strategy to search for base component descendants:<ul><li>children: return the children components of the base component. Grandchildren components are not returned</li><li>all: return all the descendants components of the base component. Grandchildren are returned.</li><li>leaves: return all the descendant components (files, in general) which don't have other children. They are the leaves of the component tree.</li></ul>",ExampleValue:""
 }
