@@ -1,5 +1,5 @@
 // Monitoring
-package sonar
+package sonargo
 
 import "net/http"
 
@@ -10,7 +10,7 @@ type MonitoringService struct {
 /*
 Metrics Return monitoring metrics in Prometheus format.
 Support content type 'text/plain' (default) and 'application/openmetrics-text'.
-this endpoint can be access using a Bearer token, that needs to be defined in sonar.properties with the 'sonar.web.systemPasscode' key.
+This endpoint can be accessed using a Bearer token, which needs to be defined in sonar.properties with the 'sonar.web.systemPasscode' key.
 */
 func (s *MonitoringService) Metrics() (v *string, resp *http.Response, err error) {
 	req, err := s.client.NewRequest("GET", "monitoring/metrics", nil)
