@@ -1,5 +1,5 @@
-// Manage new code periods.
-package sonar
+// Manage <a href="https://docs.sonarsource.com/sonarqube-community-build/project-administration/configuring-new-code-calculation#setting-specific-new-code-definition-for-project" target="_blank" rel="noopener noreferrer">new code definition</a>.
+package sonargo
 
 import "net/http"
 
@@ -31,7 +31,7 @@ type NewCodePeriodsListOption struct {
 	Project string `url:"project,omitempty"` // Description:"Project key",ExampleValue:""
 }
 
-// List List the New Code Periods for all branches in a project.<br>Requires the permission to browse the project
+// List Lists the <a href="https://docs.sonarsource.com/sonarqube-community-build/project-administration/configuring-new-code-calculation#setting-specific-new-code-definition-for-project" target="_blank" rel="noopener noreferrer">new code definition</a> for all branches in a project.<br>Requires the permission to browse the project
 func (s *NewCodePeriodsService) List(opt *NewCodePeriodsListOption) (v *NewCodePeriodsListObject, resp *http.Response, err error) {
 	err = s.ValidateListOpt(opt)
 	if err != nil {
@@ -52,11 +52,11 @@ func (s *NewCodePeriodsService) List(opt *NewCodePeriodsListOption) (v *NewCodeP
 type NewCodePeriodsSetOption struct {
 	Branch  string `url:"branch,omitempty"`  // Description:"Branch key",ExampleValue:""
 	Project string `url:"project,omitempty"` // Description:"Project key",ExampleValue:""
-	Type    string `url:"type,omitempty"`    // Description:"Type<br/>New code periods of the following types are allowed:<ul><li>SPECIFIC_ANALYSIS - can be set at branch level only</li><li>PREVIOUS_VERSION - can be set at any level (global, project, branch)</li><li>NUMBER_OF_DAYS - can be set at any level (global, project, branch)</li><li>REFERENCE_BRANCH - can only be set for projects and branches</li></ul>",ExampleValue:""
-	Value   string `url:"value,omitempty"`   // Description:"Value<br/>For each type, a different value is expected:<ul><li>the uuid of an analysis, when type is SPECIFIC_ANALYSIS</li><li>no value, when type is PREVIOUS_VERSION</li><li>a number, when type is NUMBER_OF_DAYS</li><li>a string, when type is REFERENCE_BRANCH</li></ul>",ExampleValue:""
+	Type    string `url:"type,omitempty"`    // Description:"Type<br/>New code definitions of the following types are allowed:<ul><li>SPECIFIC_ANALYSIS - can be set at branch level only</li><li>PREVIOUS_VERSION - can be set at any level (global, project, branch)</li><li>NUMBER_OF_DAYS - can be set at any level (global, project, branch)</li><li>REFERENCE_BRANCH - can only be set for projects and branches</li></ul>",ExampleValue:""
+	Value   string `url:"value,omitempty"`   // Description:"Value<br/>For each type, a different value is expected:<ul><li>the uuid of an analysis, when type is SPECIFIC_ANALYSIS</li><li>no value, when type is PREVIOUS_VERSION</li><li>a number between 1 and 90, when type is NUMBER_OF_DAYS</li><li>a string, when type is REFERENCE_BRANCH</li></ul>",ExampleValue:""
 }
 
-// Set Updates the setting for the New Code Period on different levels:<br><ul><li>Project key must be provided to update the value for a project</li><li>Both project and branch keys must be provided to update the value for a branch</li></ul>Requires one of the following permissions: <ul><li>'Administer System' to change the global setting</li><li>'Administer' rights on the specified project to change the project setting</li></ul>
+// Set Updates the <a href="https://docs.sonarsource.com/sonarqube-community-build/project-administration/configuring-new-code-calculation#setting-specific-new-code-definition-for-project" target="_blank" rel="noopener noreferrer">new code definition</a> on different levels:<br><ul><li>Not providing a project key and a branch key will update the default value at global level. Existing projects or branches having a specific new code definition will not be impacted</li><li>Project key must be provided to update the value for a project</li><li>Both project and branch keys must be provided to update the value for a branch</li></ul>Requires one of the following permissions: <ul><li>'Administer System' to change the global setting</li><li>'Administer' rights on the specified project to change the project setting</li></ul>
 func (s *NewCodePeriodsService) Set(opt *NewCodePeriodsSetOption) (resp *http.Response, err error) {
 	err = s.ValidateSetOpt(opt)
 	if err != nil {
@@ -78,7 +78,7 @@ type NewCodePeriodsShowOption struct {
 	Project string `url:"project,omitempty"` // Description:"Project key",ExampleValue:""
 }
 
-// Show Shows a setting for the New Code Period.<br> If the component requested doesn't exist or if no new code period is set for it, a value is inherited from the project or from the global setting.Requires one of the following permissions if a component is specified: <ul><li>'Administer' rights on the specified component</li><li>'Execute analysis' rights on the specified component</li></ul>
+// Show Shows the <a href="https://docs.sonarsource.com/sonarqube-community-build/project-administration/configuring-new-code-calculation#setting-specific-new-code-definition-for-project" target="_blank" rel="noopener noreferrer">new code definition</a>.<br> If the component requested doesn't exist or if no new code definition is set for it, a value is inherited from the project or from the global setting.Requires one of the following permissions if a component is specified: <ul><li>'Administer' rights on the specified component</li><li>'Execute analysis' rights on the specified component</li></ul>
 func (s *NewCodePeriodsService) Show(opt *NewCodePeriodsShowOption) (v *NewCodePeriodsShowObject, resp *http.Response, err error) {
 	err = s.ValidateShowOpt(opt)
 	if err != nil {
@@ -101,7 +101,7 @@ type NewCodePeriodsUnsetOption struct {
 	Project string `url:"project,omitempty"` // Description:"Project key",ExampleValue:""
 }
 
-// Unset Unset the New Code Period setting for a branch, project or global.<br>Requires one of the following permissions: <ul><li>'Administer System' to change the global setting</li><li>'Administer' rights for a specified component</li></ul>
+// Unset Unsets the <a href="https://docs.sonarsource.com/sonarqube-community-build/project-administration/configuring-new-code-calculation#setting-specific-new-code-definition-for-project" target="_blank" rel="noopener noreferrer">new code definition</a> for a branch, project or global. It requires the inherited New Code Definition to be compatible with the Clean as You Code methodology, and one of the following permissions: <ul><li>'Administer System' to change the global setting</li><li>'Administer' rights for a specified component</li></ul>
 func (s *NewCodePeriodsService) Unset(opt *NewCodePeriodsUnsetOption) (resp *http.Response, err error) {
 	err = s.ValidateUnsetOpt(opt)
 	if err != nil {
