@@ -1,5 +1,5 @@
 // Project export/import
-package sonar
+package sonargo
 
 import "net/http"
 
