@@ -1,5 +1,5 @@
 // Get and update some details of automatic rules, and manage custom rules.
-package sonar
+package sonargo
 
 import "net/http"
 
@@ -40,34 +40,42 @@ type RulesAppObject_sub1 struct {
 }
 
 type RulesCreateObject struct {
-	Rule RulesCreateObject_sub2 `json:"rule,omitempty"`
+	Rule RulesCreateObject_sub3 `json:"rule,omitempty"`
+}
+
+type RulesCreateObject_sub3 struct {
+	CleanCodeAttribute         string                   `json:"cleanCodeAttribute,omitempty"`
+	CleanCodeAttributeCategory string                   `json:"cleanCodeAttributeCategory,omitempty"`
+	CreatedAt                  string                   `json:"createdAt,omitempty"`
+	HTMLDesc                   string                   `json:"htmlDesc,omitempty"`
+	Impacts                    []RulesCreateObject_sub1 `json:"impacts,omitempty"`
+	IsExternal                 bool                     `json:"isExternal,omitempty"`
+	IsTemplate                 bool                     `json:"isTemplate,omitempty"`
+	Key                        string                   `json:"key,omitempty"`
+	Lang                       string                   `json:"lang,omitempty"`
+	LangName                   string                   `json:"langName,omitempty"`
+	MdDesc                     string                   `json:"mdDesc,omitempty"`
+	Name                       string                   `json:"name,omitempty"`
+	Params                     []RulesCreateObject_sub2 `json:"params,omitempty"`
+	Repo                       string                   `json:"repo,omitempty"`
+	Scope                      string                   `json:"scope,omitempty"`
+	Severity                   string                   `json:"severity,omitempty"`
+	Status                     string                   `json:"status,omitempty"`
+	SysTags                    []interface{}            `json:"sysTags,omitempty"`
+	TemplateKey                string                   `json:"templateKey,omitempty"`
+	Type                       string                   `json:"type,omitempty"`
 }
 
 type RulesCreateObject_sub2 struct {
-	CreatedAt   string                   `json:"createdAt,omitempty"`
-	HTMLDesc    string                   `json:"htmlDesc,omitempty"`
-	IsExternal  bool                     `json:"isExternal,omitempty"`
-	IsTemplate  bool                     `json:"isTemplate,omitempty"`
-	Key         string                   `json:"key,omitempty"`
-	Lang        string                   `json:"lang,omitempty"`
-	LangName    string                   `json:"langName,omitempty"`
-	MdDesc      string                   `json:"mdDesc,omitempty"`
-	Name        string                   `json:"name,omitempty"`
-	Params      []RulesCreateObject_sub1 `json:"params,omitempty"`
-	Repo        string                   `json:"repo,omitempty"`
-	Scope       string                   `json:"scope,omitempty"`
-	Severity    string                   `json:"severity,omitempty"`
-	Status      string                   `json:"status,omitempty"`
-	SysTags     []interface{}            `json:"sysTags,omitempty"`
-	TemplateKey string                   `json:"templateKey,omitempty"`
-	Type        string                   `json:"type,omitempty"`
-}
-
-type RulesCreateObject_sub1 struct {
 	DefaultValue string `json:"defaultValue,omitempty"`
 	HTMLDesc     string `json:"htmlDesc,omitempty"`
 	Key          string `json:"key,omitempty"`
 	Type         string `json:"type,omitempty"`
+}
+
+type RulesCreateObject_sub1 struct {
+	Severity        string `json:"severity,omitempty"`
+	SoftwareQuality string `json:"softwareQuality,omitempty"`
 }
 
 type RulesRepositoriesObject struct {
@@ -84,7 +92,36 @@ type RulesSearchObject struct {
 	Actives RulesSearchObject_sub3    `json:"actives,omitempty"`
 	Facets  []RulesSearchObject_sub5  `json:"facets,omitempty"`
 	Paging  RulesSearchObject_sub6    `json:"paging,omitempty"`
-	Rules   []RulesSearchObject_sub10 `json:"rules,omitempty"`
+	Rules   []RulesSearchObject_sub11 `json:"rules,omitempty"`
+}
+
+type RulesSearchObject_sub11 struct {
+	CleanCodeAttribute         string                    `json:"cleanCodeAttribute,omitempty"`
+	CleanCodeAttributeCategory string                    `json:"cleanCodeAttributeCategory,omitempty"`
+	CreatedAt                  string                    `json:"createdAt,omitempty"`
+	DescriptionSections        []RulesSearchObject_sub8  `json:"descriptionSections,omitempty"`
+	HTMLDesc                   string                    `json:"htmlDesc,omitempty"`
+	HTMLNote                   string                    `json:"htmlNote,omitempty"`
+	Impacts                    []RulesSearchObject_sub9  `json:"impacts,omitempty"`
+	InternalKey                string                    `json:"internalKey,omitempty"`
+	IsExternal                 bool                      `json:"isExternal,omitempty"`
+	IsTemplate                 bool                      `json:"isTemplate,omitempty"`
+	Key                        string                    `json:"key,omitempty"`
+	Lang                       string                    `json:"lang,omitempty"`
+	LangName                   string                    `json:"langName,omitempty"`
+	MdNote                     string                    `json:"mdNote,omitempty"`
+	Name                       string                    `json:"name,omitempty"`
+	NoteLogin                  string                    `json:"noteLogin,omitempty"`
+	Params                     []RulesSearchObject_sub10 `json:"params,omitempty"`
+	Repo                       string                    `json:"repo,omitempty"`
+	Scope                      string                    `json:"scope,omitempty"`
+	Severity                   string                    `json:"severity,omitempty"`
+	Status                     string                    `json:"status,omitempty"`
+	SysTags                    []string                  `json:"sysTags,omitempty"`
+	Tags                       []interface{}             `json:"tags,omitempty"`
+	TemplateKey                string                    `json:"templateKey,omitempty"`
+	Type                       string                    `json:"type,omitempty"`
+	UpdatedAt                  string                    `json:"updatedAt,omitempty"`
 }
 
 type RulesSearchObject_sub8 struct {
@@ -99,32 +136,6 @@ type RulesSearchObject_sub4 struct {
 }
 
 type RulesSearchObject_sub10 struct {
-	CreatedAt           string                   `json:"createdAt,omitempty"`
-	DescriptionSections []RulesSearchObject_sub8 `json:"descriptionSections,omitempty"`
-	HTMLDesc            string                   `json:"htmlDesc,omitempty"`
-	HTMLNote            string                   `json:"htmlNote,omitempty"`
-	InternalKey         string                   `json:"internalKey,omitempty"`
-	IsExternal          bool                     `json:"isExternal,omitempty"`
-	IsTemplate          bool                     `json:"isTemplate,omitempty"`
-	Key                 string                   `json:"key,omitempty"`
-	Lang                string                   `json:"lang,omitempty"`
-	LangName            string                   `json:"langName,omitempty"`
-	MdNote              string                   `json:"mdNote,omitempty"`
-	Name                string                   `json:"name,omitempty"`
-	NoteLogin           string                   `json:"noteLogin,omitempty"`
-	Params              []RulesSearchObject_sub9 `json:"params,omitempty"`
-	Repo                string                   `json:"repo,omitempty"`
-	Scope               string                   `json:"scope,omitempty"`
-	Severity            string                   `json:"severity,omitempty"`
-	Status              string                   `json:"status,omitempty"`
-	SysTags             []string                 `json:"sysTags,omitempty"`
-	Tags                []interface{}            `json:"tags,omitempty"`
-	TemplateKey         string                   `json:"templateKey,omitempty"`
-	Type                string                   `json:"type,omitempty"`
-	UpdatedAt           string                   `json:"updatedAt,omitempty"`
-}
-
-type RulesSearchObject_sub9 struct {
 	DefaultValue string `json:"defaultValue,omitempty"`
 	Desc         string `json:"desc,omitempty"`
 	Key          string `json:"key,omitempty"`
@@ -158,6 +169,11 @@ type RulesSearchObject_sub6 struct {
 	Total     int64 `json:"total,omitempty"`
 }
 
+type RulesSearchObject_sub9 struct {
+	Severity        string `json:"severity,omitempty"`
+	SoftwareQuality string `json:"softwareQuality,omitempty"`
+}
+
 type RulesSearchObject_sub3 struct {
 	Squid_ClassCyclomaticComplexity  []RulesSearchObject_sub2 `json:"squid:ClassCyclomaticComplexity,omitempty"`
 	Squid_MethodCyclomaticComplexity []RulesSearchObject_sub2 `json:"squid:MethodCyclomaticComplexity,omitempty"`
@@ -166,7 +182,38 @@ type RulesSearchObject_sub3 struct {
 
 type RulesShowObject struct {
 	Actives []RulesShowObject_sub2 `json:"actives,omitempty"`
-	Rule    RulesShowObject_sub6   `json:"rule,omitempty"`
+	Rule    RulesShowObject_sub7   `json:"rule,omitempty"`
+}
+
+type RulesShowObject_sub7 struct {
+	CleanCodeAttribute         string                 `json:"cleanCodeAttribute,omitempty"`
+	CleanCodeAttributeCategory string                 `json:"cleanCodeAttributeCategory,omitempty"`
+	DefaultRemFnBaseEffort     string                 `json:"defaultRemFnBaseEffort,omitempty"`
+	DefaultRemFnGapMultiplier  string                 `json:"defaultRemFnGapMultiplier,omitempty"`
+	DefaultRemFnType           string                 `json:"defaultRemFnType,omitempty"`
+	DescriptionSections        []RulesShowObject_sub4 `json:"descriptionSections,omitempty"`
+	GapDescription             string                 `json:"gapDescription,omitempty"`
+	HTMLDesc                   string                 `json:"htmlDesc,omitempty"`
+	Impacts                    []RulesShowObject_sub5 `json:"impacts,omitempty"`
+	InternalKey                string                 `json:"internalKey,omitempty"`
+	IsExternal                 bool                   `json:"isExternal,omitempty"`
+	Key                        string                 `json:"key,omitempty"`
+	Lang                       string                 `json:"lang,omitempty"`
+	LangName                   string                 `json:"langName,omitempty"`
+	Name                       string                 `json:"name,omitempty"`
+	Params                     []RulesShowObject_sub6 `json:"params,omitempty"`
+	RemFnBaseEffort            string                 `json:"remFnBaseEffort,omitempty"`
+	RemFnGapMultiplier         string                 `json:"remFnGapMultiplier,omitempty"`
+	RemFnOverloaded            bool                   `json:"remFnOverloaded,omitempty"`
+	RemFnType                  string                 `json:"remFnType,omitempty"`
+	Repo                       string                 `json:"repo,omitempty"`
+	Scope                      string                 `json:"scope,omitempty"`
+	Severity                   string                 `json:"severity,omitempty"`
+	Status                     string                 `json:"status,omitempty"`
+	SysTags                    []string               `json:"sysTags,omitempty"`
+	Tags                       []interface{}          `json:"tags,omitempty"`
+	Template                   bool                   `json:"template,omitempty"`
+	Type                       string                 `json:"type,omitempty"`
 }
 
 type RulesShowObject_sub4 struct {
@@ -176,34 +223,6 @@ type RulesShowObject_sub4 struct {
 }
 
 type RulesShowObject_sub6 struct {
-	DefaultRemFnBaseEffort    string                 `json:"defaultRemFnBaseEffort,omitempty"`
-	DefaultRemFnGapMultiplier string                 `json:"defaultRemFnGapMultiplier,omitempty"`
-	DefaultRemFnType          string                 `json:"defaultRemFnType,omitempty"`
-	DescriptionSections       []RulesShowObject_sub4 `json:"descriptionSections,omitempty"`
-	GapDescription            string                 `json:"gapDescription,omitempty"`
-	HTMLDesc                  string                 `json:"htmlDesc,omitempty"`
-	InternalKey               string                 `json:"internalKey,omitempty"`
-	IsExternal                bool                   `json:"isExternal,omitempty"`
-	Key                       string                 `json:"key,omitempty"`
-	Lang                      string                 `json:"lang,omitempty"`
-	LangName                  string                 `json:"langName,omitempty"`
-	Name                      string                 `json:"name,omitempty"`
-	Params                    []RulesShowObject_sub5 `json:"params,omitempty"`
-	RemFnBaseEffort           string                 `json:"remFnBaseEffort,omitempty"`
-	RemFnGapMultiplier        string                 `json:"remFnGapMultiplier,omitempty"`
-	RemFnOverloaded           bool                   `json:"remFnOverloaded,omitempty"`
-	RemFnType                 string                 `json:"remFnType,omitempty"`
-	Repo                      string                 `json:"repo,omitempty"`
-	Scope                     string                 `json:"scope,omitempty"`
-	Severity                  string                 `json:"severity,omitempty"`
-	Status                    string                 `json:"status,omitempty"`
-	SysTags                   []string               `json:"sysTags,omitempty"`
-	Tags                      []interface{}          `json:"tags,omitempty"`
-	Template                  bool                   `json:"template,omitempty"`
-	Type                      string                 `json:"type,omitempty"`
-}
-
-type RulesShowObject_sub5 struct {
 	DefaultValue string `json:"defaultValue,omitempty"`
 	Desc         string `json:"desc,omitempty"`
 	Key          string `json:"key,omitempty"`
@@ -215,10 +234,11 @@ type RulesShowObject_sub3 struct {
 }
 
 type RulesShowObject_sub2 struct {
-	Inherit  string                 `json:"inherit,omitempty"`
-	Params   []RulesShowObject_sub1 `json:"params,omitempty"`
-	QProfile string                 `json:"qProfile,omitempty"`
-	Severity string                 `json:"severity,omitempty"`
+	Inherit         string                 `json:"inherit,omitempty"`
+	Params          []RulesShowObject_sub1 `json:"params,omitempty"`
+	PrioritizedRule bool                   `json:"prioritizedRule,omitempty"`
+	QProfile        string                 `json:"qProfile,omitempty"`
+	Severity        string                 `json:"severity,omitempty"`
 }
 
 type RulesShowObject_sub1 struct {
@@ -226,42 +246,54 @@ type RulesShowObject_sub1 struct {
 	Value string `json:"value,omitempty"`
 }
 
+type RulesShowObject_sub5 struct {
+	Severity        string `json:"severity,omitempty"`
+	SoftwareQuality string `json:"softwareQuality,omitempty"`
+}
+
 type RulesTagsObject struct {
 	Tags []string `json:"tags,omitempty"`
 }
 
 type RulesUpdateObject struct {
-	Rule RulesUpdateObject_sub2 `json:"rule,omitempty"`
+	Rule RulesUpdateObject_sub3 `json:"rule,omitempty"`
+}
+
+type RulesUpdateObject_sub3 struct {
+	CleanCodeAttribute         string                   `json:"cleanCodeAttribute,omitempty"`
+	CleanCodeAttributeCategory string                   `json:"cleanCodeAttributeCategory,omitempty"`
+	CreatedAt                  string                   `json:"createdAt,omitempty"`
+	HTMLDesc                   string                   `json:"htmlDesc,omitempty"`
+	Impacts                    []RulesUpdateObject_sub1 `json:"impacts,omitempty"`
+	IsExternal                 bool                     `json:"isExternal,omitempty"`
+	IsTemplate                 bool                     `json:"isTemplate,omitempty"`
+	Key                        string                   `json:"key,omitempty"`
+	Lang                       string                   `json:"lang,omitempty"`
+	LangName                   string                   `json:"langName,omitempty"`
+	MdDesc                     string                   `json:"mdDesc,omitempty"`
+	Name                       string                   `json:"name,omitempty"`
+	Params                     []RulesUpdateObject_sub2 `json:"params,omitempty"`
+	RemFnOverloaded            bool                     `json:"remFnOverloaded,omitempty"`
+	Repo                       string                   `json:"repo,omitempty"`
+	Scope                      string                   `json:"scope,omitempty"`
+	Severity                   string                   `json:"severity,omitempty"`
+	Status                     string                   `json:"status,omitempty"`
+	SysTags                    []interface{}            `json:"sysTags,omitempty"`
+	Tags                       []interface{}            `json:"tags,omitempty"`
+	TemplateKey                string                   `json:"templateKey,omitempty"`
+	Type                       string                   `json:"type,omitempty"`
 }
 
 type RulesUpdateObject_sub2 struct {
-	CreatedAt       string                   `json:"createdAt,omitempty"`
-	DebtOverloaded  bool                     `json:"debtOverloaded,omitempty"`
-	HTMLDesc        string                   `json:"htmlDesc,omitempty"`
-	IsExternal      bool                     `json:"isExternal,omitempty"`
-	IsTemplate      bool                     `json:"isTemplate,omitempty"`
-	Key             string                   `json:"key,omitempty"`
-	Lang            string                   `json:"lang,omitempty"`
-	LangName        string                   `json:"langName,omitempty"`
-	MdDesc          string                   `json:"mdDesc,omitempty"`
-	Name            string                   `json:"name,omitempty"`
-	Params          []RulesUpdateObject_sub1 `json:"params,omitempty"`
-	RemFnOverloaded bool                     `json:"remFnOverloaded,omitempty"`
-	Repo            string                   `json:"repo,omitempty"`
-	Scope           string                   `json:"scope,omitempty"`
-	Severity        string                   `json:"severity,omitempty"`
-	Status          string                   `json:"status,omitempty"`
-	SysTags         []interface{}            `json:"sysTags,omitempty"`
-	Tags            []interface{}            `json:"tags,omitempty"`
-	TemplateKey     string                   `json:"templateKey,omitempty"`
-	Type            string                   `json:"type,omitempty"`
-}
-
-type RulesUpdateObject_sub1 struct {
 	DefaultValue string `json:"defaultValue,omitempty"`
 	HTMLDesc     string `json:"htmlDesc,omitempty"`
 	Key          string `json:"key,omitempty"`
 	Type         string `json:"type,omitempty"`
+}
+
+type RulesUpdateObject_sub1 struct {
+	Severity        string `json:"severity,omitempty"`
+	SoftwareQuality string `json:"softwareQuality,omitempty"`
 }
 
 // App Get data required for rendering the page 'Coding Rules'.
@@ -279,14 +311,16 @@ func (s *RulesService) App() (v *RulesAppObject, resp *http.Response, err error)
 }
 
 type RulesCreateOption struct {
+	CleanCodeAttribute  string `url:"cleanCodeAttribute,omitempty"`  // Description:"Clean code attribute",ExampleValue:""
 	CustomKey           string `url:"customKey,omitempty"`           // Description:"Key of the custom rule",ExampleValue:"Todo_should_not_be_used"
+	Impacts             string `url:"impacts,omitempty"`             // Description:"Impacts as semi-colon list of &lt;software_quality&gt;=&lt;severity&gt;",ExampleValue:"SECURITY=HIGH;MAINTAINABILITY=LOW"
 	MarkdownDescription string `url:"markdownDescription,omitempty"` // Description:"Rule description in <a href='/formatting/help'>markdown format</a>",ExampleValue:"Description of my custom rule"
 	Name                string `url:"name,omitempty"`                // Description:"Rule name",ExampleValue:"My custom rule"
-	Params              string `url:"params,omitempty"`              // Description:"Parameters as semi-colon list of <key>=<value>, for example 'params=key1=v1;key2=v2' (Only for custom rule)",ExampleValue:""
+	Params              string `url:"params,omitempty"`              // Description:"Parameters as semi-colon list of &lt;key&gt;=&lt;value&gt;",ExampleValue:"key1=v1;key2=v2"
 	PreventReactivation string `url:"preventReactivation,omitempty"` // Description:"If set to true and if the rule has been deactivated (status 'REMOVED'), a status 409 will be returned",ExampleValue:""
 	Severity            string `url:"severity,omitempty"`            // Description:"Rule severity",ExampleValue:""
 	Status              string `url:"status,omitempty"`              // Description:"Rule status",ExampleValue:""
-	TemplateKey         string `url:"templateKey,omitempty"`         // Description:"Key of the template rule in order to create a custom rule (mandatory for custom rule)",ExampleValue:"java:XPath"
+	TemplateKey         string `url:"templateKey,omitempty"`         // Description:"Key of the template rule in order to create a custom rule",ExampleValue:"java:XPath"
 	Type                string `url:"type,omitempty"`                // Description:"Rule type",ExampleValue:""
 }
 
@@ -329,9 +363,22 @@ func (s *RulesService) Delete(opt *RulesDeleteOption) (resp *http.Response, err 
 	return
 }
 
-// List List rules, excluding the manual rules and the rules with status REMOVED. JSON format is not supported for response.
-func (s *RulesService) List() (v *string, resp *http.Response, err error) {
-	req, err := s.client.NewRequest("GET", "rules/list", nil)
+type RulesListOption struct {
+	Asc            string `url:"asc,omitempty"`             // Description:"Ascending sort",ExampleValue:""
+	AvailableSince string `url:"available_since,omitempty"` // Description:"Filter rules available since the given date. If no value is provided, all rules are returned. Format is yyyy-MM-dd.",ExampleValue:"2014-06-22"
+	P              string `url:"p,omitempty"`               // Description:"1-based page number",ExampleValue:"42"
+	Ps             string `url:"ps,omitempty"`              // Description:"Page size. Must be greater than 0 and less than 500",ExampleValue:"20"
+	Qprofile       string `url:"qprofile,omitempty"`        // Description:"Filter rules that are activated in the given quality profile.",ExampleValue:""
+	S              string `url:"s,omitempty"`               // Description:"Sort field",ExampleValue:""
+}
+
+// List List rules, excluding the external rules and the rules with status REMOVED. If a quality profile is provided, it returns only the rules activated in this quality profile. If no quality profile is provided, it returns all the rules available to the user.
+func (s *RulesService) List(opt *RulesListOption) (v *string, resp *http.Response, err error) {
+	err = s.ValidateListOpt(opt)
+	if err != nil {
+		return
+	}
+	req, err := s.client.NewRequest("GET", "rules/list", opt)
 	if err != nil {
 		return
 	}
@@ -367,34 +414,41 @@ func (s *RulesService) Repositories(opt *RulesRepositoriesOption) (v *RulesRepos
 }
 
 type RulesSearchOption struct {
-	Activation          string `url:"activation,omitempty"`          // Description:"Filter rules that are activated or deactivated on the selected Quality profile. Ignored if the parameter 'qprofile' is not set.",ExampleValue:""
-	ActiveSeverities    string `url:"active_severities,omitempty"`   // Description:"Comma-separated list of activation severities, i.e the severity of rules in Quality profiles.",ExampleValue:"CRITICAL,BLOCKER"
-	Asc                 string `url:"asc,omitempty"`                 // Description:"Ascending sort",ExampleValue:""
-	AvailableSince      string `url:"available_since,omitempty"`     // Description:"Filters rules added since date. Format is yyyy-MM-dd",ExampleValue:"2014-06-22"
-	CompareToProfile    string `url:"compareToProfile,omitempty"`    // Description:"Quality profile key to filter rules that are activated. Meant to compare easily to profile set in 'qprofile'",ExampleValue:"AU-TpxcA-iU5OvuD2FLz"
-	Cwe                 string `url:"cwe,omitempty"`                 // Description:"Comma-separated list of CWE identifiers. Use 'unknown' to select rules not associated to any CWE.",ExampleValue:"12,125,unknown"
-	F                   string `url:"f,omitempty"`                   // Description:"Comma-separated list of additional fields to be returned in the response. All the fields are returned by default, except actives.",ExampleValue:"effortToFixDescription,defaultDebtRemFn"
-	Facets              string `url:"facets,omitempty"`              // Description:"Comma-separated list of the facets to be computed. No facet is computed by default.",ExampleValue:"languages,repositories"
-	IncludeExternal     string `url:"include_external,omitempty"`    // Description:"Include external engine rules in the results",ExampleValue:""
-	Inheritance         string `url:"inheritance,omitempty"`         // Description:"Comma-separated list of values of inheritance for a rule within a quality profile. Used only if the parameter 'activation' is set.",ExampleValue:"INHERITED,OVERRIDES"
-	IsTemplate          string `url:"is_template,omitempty"`         // Description:"Filter template rules",ExampleValue:""
-	Languages           string `url:"languages,omitempty"`           // Description:"Comma-separated list of languages",ExampleValue:"java,js"
-	OwaspTop10          string `url:"owaspTop10,omitempty"`          // Description:"Comma-separated list of OWASP Top 10 2017 lowercase categories.",ExampleValue:""
-	OwaspTop102021      string `url:"owaspTop10-2021,omitempty"`     // Description:"Comma-separated list of OWASP Top 10 2021 lowercase categories.",ExampleValue:""
-	P                   string `url:"p,omitempty"`                   // Description:"1-based page number",ExampleValue:"42"
-	Ps                  string `url:"ps,omitempty"`                  // Description:"Page size. Must be greater than 0 and less or equal than 500",ExampleValue:"20"
-	Q                   string `url:"q,omitempty"`                   // Description:"UTF-8 search query",ExampleValue:"xpath"
-	Qprofile            string `url:"qprofile,omitempty"`            // Description:"Quality profile key to filter on. Used only if the parameter 'activation' is set.",ExampleValue:"AU-Tpxb--iU5OvuD2FLy"
-	Repositories        string `url:"repositories,omitempty"`        // Description:"Comma-separated list of repositories",ExampleValue:"java,html"
-	RuleKey             string `url:"rule_key,omitempty"`            // Description:"Key of rule to search for",ExampleValue:"java:S1144"
-	S                   string `url:"s,omitempty"`                   // Description:"Sort field",ExampleValue:"name"
-	SansTop25           string `url:"sansTop25,omitempty"`           // Description:"Comma-separated list of SANS Top 25 categories.",ExampleValue:""
-	Severities          string `url:"severities,omitempty"`          // Description:"Comma-separated list of default severities. Not the same than severity of rules in Quality profiles.",ExampleValue:"CRITICAL,BLOCKER"
-	SonarsourceSecurity string `url:"sonarsourceSecurity,omitempty"` // Description:"Comma-separated list of SonarSource security categories. Use 'others' to select rules not associated with any category",ExampleValue:"sql-injection,command-injection,others"
-	Statuses            string `url:"statuses,omitempty"`            // Description:"Comma-separated list of status codes",ExampleValue:"READY"
-	Tags                string `url:"tags,omitempty"`                // Description:"Comma-separated list of tags. Returned rules match any of the tags (OR operator)",ExampleValue:"security,java8"
-	TemplateKey         string `url:"template_key,omitempty"`        // Description:"Key of the template rule to filter on. Used to search for the custom rules based on this template.",ExampleValue:"java:S001"
-	Types               string `url:"types,omitempty"`               // Description:"Comma-separated list of types. Returned rules match any of the tags (OR operator)",ExampleValue:"BUG"
+	Activation                   string `url:"activation,omitempty"`                   // Description:"Filter rules that are activated or deactivated on the selected Quality profile. Ignored if the parameter 'qprofile' is not set.",ExampleValue:""
+	ActiveImpactSeverities       string `url:"active_impactSeverities,omitempty"`      // Description:"Comma-separated list of Activation Software Quality Severities, i.e the impact severity of rules in Quality profiles.",ExampleValue:"HIGH,MEDIUM"
+	ActiveSeverities             string `url:"active_severities,omitempty"`            // Description:"Comma-separated list of activation severities, i.e the severity of rules in Quality profiles.",ExampleValue:"CRITICAL,BLOCKER"
+	Asc                          string `url:"asc,omitempty"`                          // Description:"Ascending sort",ExampleValue:""
+	AvailableSince               string `url:"available_since,omitempty"`              // Description:"Filters rules added since date. Format is yyyy-MM-dd",ExampleValue:"2014-06-22"
+	CleanCodeAttributeCategories string `url:"cleanCodeAttributeCategories,omitempty"` // Description:"Comma-separated list of Clean Code Attribute Categories",ExampleValue:"ADAPTABLE,INTENTIONAL"
+	CompareToProfile             string `url:"compareToProfile,omitempty"`             // Description:"Quality profile key to filter rules that are activated. Meant to compare easily to profile set in 'qprofile'",ExampleValue:"AU-TpxcA-iU5OvuD2FLz"
+	ComplianceStandards          string `url:"complianceStandards,omitempty"`          // Description:"Comma-separated list of compliance standards",ExampleValue:"asvs:4.0.3=6.1.2"
+	Cwe                          string `url:"cwe,omitempty"`                          // Description:"Comma-separated list of CWE identifiers. Use 'unknown' to select rules not associated to any CWE.",ExampleValue:"12,125,unknown"
+	F                            string `url:"f,omitempty"`                            // Description:"Comma-separated list of additional fields to be returned in the response. All the fields are returned by default, except actives.",ExampleValue:"isTemplate,templateKey"
+	Facets                       string `url:"facets,omitempty"`                       // Description:"Comma-separated list of the facets to be computed. No facet is computed by default.",ExampleValue:"languages,repositories"
+	ImpactSeverities             string `url:"impactSeverities,omitempty"`             // Description:"Comma-separated list of Software Quality Severities",ExampleValue:"HIGH,MEDIUM"
+	ImpactSoftwareQualities      string `url:"impactSoftwareQualities,omitempty"`      // Description:"Comma-separated list of Software Qualities",ExampleValue:"MAINTAINABILITY,RELIABILITY"
+	IncludeExternal              string `url:"include_external,omitempty"`             // Description:"Include external engine rules in the results",ExampleValue:""
+	Inheritance                  string `url:"inheritance,omitempty"`                  // Description:"Comma-separated list of values of inheritance for a rule within a quality profile. Used only if the parameter 'activation' is set.",ExampleValue:"INHERITED,OVERRIDES"
+	IsTemplate                   string `url:"is_template,omitempty"`                  // Description:"Filter template rules",ExampleValue:""
+	Languages                    string `url:"languages,omitempty"`                    // Description:"Comma-separated list of languages",ExampleValue:"java,js"
+	OwaspMobileTop102024         string `url:"owaspMobileTop10-2024,omitempty"`        // Description:"Comma-separated list of OWASP Mobile Top 10 2024 lowercase categories.",ExampleValue:""
+	OwaspTop10                   string `url:"owaspTop10,omitempty"`                   // Description:"Comma-separated list of OWASP Top 10 2017 lowercase categories.",ExampleValue:""
+	OwaspTop102021               string `url:"owaspTop10-2021,omitempty"`              // Description:"Comma-separated list of OWASP Top 10 2021 lowercase categories.",ExampleValue:""
+	P                            string `url:"p,omitempty"`                            // Description:"1-based page number",ExampleValue:"42"
+	PrioritizedRule              string `url:"prioritizedRule,omitempty"`              // Description:"Filter on prioritized rules. Ignored if the parameter 'qprofile' is not set.",ExampleValue:""
+	Ps                           string `url:"ps,omitempty"`                           // Description:"Page size. Must be greater than 0 and less or equal than 500",ExampleValue:"20"
+	Q                            string `url:"q,omitempty"`                            // Description:"UTF-8 search query",ExampleValue:"xpath"
+	Qprofile                     string `url:"qprofile,omitempty"`                     // Description:"Quality profile key to filter on. Only rules of the same language as this profile are returned. By default only rules activated in this profile are returned. You can change that using the 'activation' parameter.",ExampleValue:"AU-Tpxb--iU5OvuD2FLy"
+	Repositories                 string `url:"repositories,omitempty"`                 // Description:"Comma-separated list of repositories",ExampleValue:"java,html"
+	RuleKey                      string `url:"rule_key,omitempty"`                     // Description:"Key of rule to search for",ExampleValue:"java:S1144"
+	S                            string `url:"s,omitempty"`                            // Description:"Sort field",ExampleValue:"name"
+	SansTop25                    string `url:"sansTop25,omitempty"`                    // Description:"Comma-separated list of SANS Top 25 categories.",ExampleValue:""
+	Severities                   string `url:"severities,omitempty"`                   // Description:"Comma-separated list of default severities. Not the same than severity of rules in Quality profiles.",ExampleValue:"CRITICAL,BLOCKER"
+	SonarsourceSecurity          string `url:"sonarsourceSecurity,omitempty"`          // Description:"Comma-separated list of SonarSource security categories. Use 'others' to select rules not associated with any category",ExampleValue:"sql-injection,command-injection,others"
+	Statuses                     string `url:"statuses,omitempty"`                     // Description:"Comma-separated list of status codes",ExampleValue:"READY"
+	Tags                         string `url:"tags,omitempty"`                         // Description:"Comma-separated list of tags. Returned rules match any of the tags (OR operator)",ExampleValue:"security,java8"
+	TemplateKey                  string `url:"template_key,omitempty"`                 // Description:"Key of the template rule to filter on. Used to search for the custom rules based on this template.",ExampleValue:"java:S001"
+	Types                        string `url:"types,omitempty"`                        // Description:"Comma-separated list of types. Returned rules match any of the tags (OR operator)",ExampleValue:"BUG"
 }
 
 // Search Search for a collection of relevant rules matching a specified query.<br/>
@@ -462,8 +516,9 @@ func (s *RulesService) Tags(opt *RulesTagsOption) (v *RulesTagsObject, resp *htt
 }
 
 type RulesUpdateOption struct {
+	Impacts                    string `url:"impacts,omitempty"`                       // Description:"Rule impacts, semicolon-separated (Only when updating a custom rule impact severity)",ExampleValue:"MAINTAINABILITY=HIGH;SECURITY=LOW"
 	Key                        string `url:"key,omitempty"`                           // Description:"Key of the rule to update",ExampleValue:"javascript:NullCheck"
-	MarkdownDescription        string `url:"markdown_description,omitempty"`          // Description:"Rule description (mandatory for custom rule and manual rule) in <a href='/formatting/help'>markdown format</a>",ExampleValue:"Description of my custom rule"
+	MarkdownDescription        string `url:"markdownDescription,omitempty"`           // Description:"Rule description (mandatory for custom rule and manual rule) in <a href='/formatting/help'>markdown format</a>",ExampleValue:"Description of my custom rule"
 	MarkdownNote               string `url:"markdown_note,omitempty"`                 // Description:"Optional note in <a href='/formatting/help'>markdown format</a>. Use empty value to remove current note. Note is not changed if the parameter is not set.",ExampleValue:"my *note*"
 	Name                       string `url:"name,omitempty"`                          // Description:"Rule name (mandatory for custom rule)",ExampleValue:"My custom rule"
 	Params                     string `url:"params,omitempty"`                        // Description:"Parameters as semi-colon list of <key>=<value>, for example 'params=key1=v1;key2=v2' (Only when updating a custom rule)",ExampleValue:""
