@@ -1,6 +1,6 @@
-// [TODO] Code generated by "github.com/magicsong/generate-go-for-sonarqube", you should edit some validation logic here
+// [TODO] Code generated by "github.com/boxboxjason/sonarqube-client-go", you should edit some validation logic here
 
-package sonar
+package sonargo
 
 func (s *AlmIntegrationsService) ValidateCheckPatOpt(opt *AlmIntegrationsCheckPatOption) error {
 	return nil
@@ -71,28 +71,10 @@ func (s *AlmSettingsService) ValidateCreateGitlabOpt(opt *AlmSettingsCreateGitla
 func (s *AlmSettingsService) ValidateDeleteOpt(opt *AlmSettingsDeleteOption) error {
 	return nil
 }
-func (s *AlmSettingsService) ValidateDeleteBindingOpt(opt *AlmSettingsDeleteBindingOption) error {
-	return nil
-}
 func (s *AlmSettingsService) ValidateGetBindingOpt(opt *AlmSettingsGetBindingOption) error {
 	return nil
 }
 func (s *AlmSettingsService) ValidateListOpt(opt *AlmSettingsListOption) error {
-	return nil
-}
-func (s *AlmSettingsService) ValidateSetAzureBindingOpt(opt *AlmSettingsSetAzureBindingOption) error {
-	return nil
-}
-func (s *AlmSettingsService) ValidateSetBitbucketBindingOpt(opt *AlmSettingsSetBitbucketBindingOption) error {
-	return nil
-}
-func (s *AlmSettingsService) ValidateSetBitbucketcloudBindingOpt(opt *AlmSettingsSetBitbucketcloudBindingOption) error {
-	return nil
-}
-func (s *AlmSettingsService) ValidateSetGithubBindingOpt(opt *AlmSettingsSetGithubBindingOption) error {
-	return nil
-}
-func (s *AlmSettingsService) ValidateSetGitlabBindingOpt(opt *AlmSettingsSetGitlabBindingOption) error {
 	return nil
 }
 func (s *AlmSettingsService) ValidateUpdateAzureOpt(opt *AlmSettingsUpdateAzureOption) error {
@@ -111,9 +93,6 @@ func (s *AlmSettingsService) ValidateUpdateGitlabOpt(opt *AlmSettingsUpdateGitla
 	return nil
 }
 func (s *AlmSettingsService) ValidateValidateOpt(opt *AlmSettingsValidateOption) error {
-	return nil
-}
-func (s *AlmSettingsService) ValidateValidateBindingOpt(opt *AlmSettingsValidateBindingOption) error {
 	return nil
 }
 func (s *AnalysisCacheService) ValidateClearOpt(opt *AnalysisCacheClearOption) error {
@@ -170,6 +149,12 @@ func (s *ComponentsService) ValidateTreeOpt(opt *ComponentsTreeOption) error {
 func (s *DevelopersService) ValidateSearchEventsOpt(opt *DevelopersSearchEventsOption) error {
 	return nil
 }
+func (s *DismissMessageService) ValidateCheckOpt(opt *DismissMessageCheckOption) error {
+	return nil
+}
+func (s *DismissMessageService) ValidateDismissOpt(opt *DismissMessageDismissOption) error {
+	return nil
+}
 func (s *DuplicationsService) ValidateShowOpt(opt *DuplicationsShowOption) error {
 	return nil
 }
@@ -200,6 +185,12 @@ func (s *HotspotsService) ValidateDeleteCommentOpt(opt *HotspotsDeleteCommentOpt
 func (s *HotspotsService) ValidateEditCommentOpt(opt *HotspotsEditCommentOption) error {
 	return nil
 }
+func (s *HotspotsService) ValidateListOpt(opt *HotspotsListOption) error {
+	return nil
+}
+func (s *HotspotsService) ValidatePullOpt(opt *HotspotsPullOption) error {
+	return nil
+}
 func (s *HotspotsService) ValidateSearchOpt(opt *HotspotsSearchOption) error {
 	return nil
 }
@@ -207,6 +198,9 @@ func (s *HotspotsService) ValidateShowOpt(opt *HotspotsShowOption) error {
 	return nil
 }
 func (s *IssuesService) ValidateAddCommentOpt(opt *IssuesAddCommentOption) error {
+	return nil
+}
+func (s *IssuesService) ValidateAnticipatedTransitionsOpt(opt *IssuesAnticipatedTransitionsOption) error {
 	return nil
 }
 func (s *IssuesService) ValidateAssignOpt(opt *IssuesAssignOption) error {
@@ -231,6 +225,9 @@ func (s *IssuesService) ValidateDoTransitionOpt(opt *IssuesDoTransitionOption) e
 	return nil
 }
 func (s *IssuesService) ValidateEditCommentOpt(opt *IssuesEditCommentOption) error {
+	return nil
+}
+func (s *IssuesService) ValidateListOpt(opt *IssuesListOption) error {
 	return nil
 }
 func (s *IssuesService) ValidatePullOpt(opt *IssuesPullOption) error {
@@ -392,12 +389,6 @@ func (s *ProjectAnalysesService) ValidateDeleteEventOpt(opt *ProjectAnalysesDele
 func (s *ProjectAnalysesService) ValidateSearchOpt(opt *ProjectAnalysesSearchOption) error {
 	return nil
 }
-func (s *ProjectAnalysesService) ValidateSetBaselineOpt(opt *ProjectAnalysesSetBaselineOption) error {
-	return nil
-}
-func (s *ProjectAnalysesService) ValidateUnsetBaselineOpt(opt *ProjectAnalysesUnsetBaselineOption) error {
-	return nil
-}
 func (s *ProjectAnalysesService) ValidateUpdateEventOpt(opt *ProjectAnalysesUpdateEventOption) error {
 	return nil
 }
@@ -425,6 +416,9 @@ func (s *ProjectBranchesService) ValidateRenameOpt(opt *ProjectBranchesRenameOpt
 func (s *ProjectBranchesService) ValidateSetAutomaticDeletionProtectionOpt(opt *ProjectBranchesSetAutomaticDeletionProtectionOption) error {
 	return nil
 }
+func (s *ProjectBranchesService) ValidateSetMainOpt(opt *ProjectBranchesSetMainOption) error {
+	return nil
+}
 func (s *ProjectDumpService) ValidateExportOpt(opt *ProjectDumpExportOption) error {
 	return nil
 }
@@ -438,12 +432,6 @@ func (s *ProjectLinksService) ValidateDeleteOpt(opt *ProjectLinksDeleteOption) e
 	return nil
 }
 func (s *ProjectLinksService) ValidateSearchOpt(opt *ProjectLinksSearchOption) error {
-	return nil
-}
-func (s *ProjectPullRequestsService) ValidateDeleteOpt(opt *ProjectPullRequestsDeleteOption) error {
-	return nil
-}
-func (s *ProjectPullRequestsService) ValidateListOpt(opt *ProjectPullRequestsListOption) error {
 	return nil
 }
 func (s *ProjectTagsService) ValidateSearchOpt(opt *ProjectTagsSearchOption) error {
@@ -629,6 +617,9 @@ func (s *RulesService) ValidateCreateOpt(opt *RulesCreateOption) error {
 func (s *RulesService) ValidateDeleteOpt(opt *RulesDeleteOption) error {
 	return nil
 }
+func (s *RulesService) ValidateListOpt(opt *RulesListOption) error {
+	return nil
+}
 func (s *RulesService) ValidateRepositoriesOpt(opt *RulesRepositoriesOption) error {
 	return nil
 }
@@ -768,5 +759,11 @@ func (s *WebservicesService) ValidateListOpt(opt *WebservicesListOption) error {
 	return nil
 }
 func (s *WebservicesService) ValidateResponseExampleOpt(opt *WebservicesResponseExampleOption) error {
+	return nil
+}
+func (s *HService) ValidateFileOpt(opt *HFileOption) error {
+	return nil
+}
+func (s *HService) ValidateProjectOpt(opt *HProjectOption) error {
 	return nil
 }
